+++ src/Psr16/InvalidArgumentException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace MatthiasMullie\Scrapbook\Psr16;
+
+use MatthiasMullie\Scrapbook\Exception\Exception;
+
+class InvalidArgumentException extends Exception implements \Psr\SimpleCache\InvalidArgumentException
+{
+}